+++ CHANGELOG.rst
@@ -1,9 +1,18 @@
 Changelog
 =========
 
+Version 0.1.2
+-------------
+- Make 2 direction builders available in the `atlas-direction-vectors direction-vectors` application:
+    * layer-region: can specify for any layered region what the values are for
+      certain layer voxels, such that the blur-gradient is more accurate
+    * source-target-layered-region:
+
+
+Version 0.1.1
+-------------
+- Open source release
+
 Version 0.1.0
 -------------
-- Initial commit: extracts the `direction_vectors` module from `nse/atlas-building-tools` [`NSETM_1685_`]
-
-
-.. _`NSETM-1685`: https://bbpteam.epfl.ch/project/issues/browse/NSETM-1685
+- Initial commit: extracts the `direction_vectors` from internal tools